added simple calculation of distance and price
